requirements.txt
asgiref==3.8.1
astroid==3.3.5
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
dill==0.3.9
Django==5.1.1
django-extensions==3.2.3
django-multiselectfield==0.1.13
executing==2.1.0
ipython==8.27.0
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomlkit==0.13.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13

django-debug-toolbar~=4.4.6